Move FramePointerValidator to its own module
This is currently in OrbitCore but the rest of OrbitCore doesn't need it, only
OrbitService.
