+++ OrbitFramePointerValidator/FunctionFramePointerValidator.cpp
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "FunctionFramePointerValidator.h"
+#include "include/OrbitFramePointerValidator/FunctionFramePointerValidator.h"
 
 #include <capstone/capstone.h>
