+++ OrbitService/CMakeLists.txt
@@ -44,6 +44,7 @@ target_include_directories(OrbitServiceLib PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 
 target_link_libraries(OrbitServiceLib PUBLIC
         OrbitCore
+        OrbitFramePointerValidator
         OrbitProtos)
 
 project(OrbitService)
